refactor: move job classes (#63)
