+++ tests/job.test.ts
@@ -3,7 +3,7 @@ import { cycleMigrations } from '../src/pg/migrations';
 import { PgStore } from '../src/pg/pg-store';
 import { DbJob, DbSipNumber, DbSmartContractInsert } from '../src/pg/types';
 import { RetryableJobError } from '../src/token-processor/queue/errors';
-import { Job } from '../src/token-processor/queue/job';
+import { Job } from '../src/token-processor/queue/job/job';
 
 class TestRetryableJob extends Job {
   description(): string {
