+++ tests/process-smart-contract-job.test.ts
@@ -2,7 +2,7 @@ import { cvToHex, uintCV } from '@stacks/transactions';
 import { MockAgent, setGlobalDispatcher } from 'undici';
 import { PgStore } from '../src/pg/pg-store';
 import { DbSipNumber, DbSmartContractInsert, DbToken, DbTokenType } from '../src/pg/types';
-import { ProcessSmartContractJob } from '../src/token-processor/process-smart-contract-job';
+import { ProcessSmartContractJob } from '../src/token-processor/queue/job/process-smart-contract-job';
 import { ENV } from '../src/env';
 import { cycleMigrations } from '../src/pg/migrations';
